Plotting part 2
data frame EU: contains data on the per-capita GDP in PPS (artificial currency unit). Each row of the data frame corresponds to a year and each column corresponds to a country.

a) Create a plot of the per-capita GDP in the UK from 1995 to 2010 with labeled axis and title.
b) Create a plot that shows the data for all countries with different colours. Add a legend. Make sure all lines are fully visible. 
c) Create a data frame EU.normalised which contains the normalised data (by GDP in year 2000).
d) Create a plot with EU.normalised and draw a dashed vertical line corresponding to the year 2000.